create custome hook for client socket connection to simplify component logic
